requirements.txt
Flask==2.0.1
twilio==7.17.0
python-dotenv==0.19.0
email-validator==1.1.3
gunicorn==20.1.0
